+++ pytestqt/qtbot.py
@@ -435,11 +435,16 @@ class QtBot(object):
         blocker.wait()
 
     @contextlib.contextmanager
-    def assertNotEmitted(self, signal):
+    def assertNotEmitted(self, signal, wait=0):
         """
         .. versionadded:: 1.11
 
         Make sure the given ``signal`` doesn't get emitted.
+
+        :param int wait:
+            How many milliseconds to wait to make sure the signal isn't emitted
+            asynchronously. By default, this method returns immediately and only
+            catches signals emitted inside the ``with``-block.
 
         This is intended to be used as a context manager.
 
@@ -447,7 +452,7 @@ class QtBot(object):
                   (pep-8 alias)
         """
         spy = SignalEmittedSpy(signal)
-        with spy:
+        with spy, self.waitSignal(signal, timeout=wait, raising=False):
             yield
         spy.assert_not_emitted()